Fix parsing of github response
The response code of the curl call seems to be dependent on the version
of `curl` used, and which exact protocol is used.

On curl 8.14.1 (Windows) I see `HTTP/1.1 200 OK`, but on curl 8.7.1
(x86_64-apple-darwin24.0) I get `HTTP/2 200 `.

This grep pattern is robust to both, and hopefully will continue to be
so.